♻️ refactor(server): extract main.rs functions into dedicated modules (#92)
## Why (User Story)
As a developer, I want main.rs to focus on server setup and routing so that
business logic lives in dedicated, testable modules.

## What Changed
- Extracted 1037 lines from main.rs into 5 new modules:
  - background.rs: Segment loading, graph building, catalog persistence
  - credential_store.rs: Secure credential storage with encryption
  - export.rs: S3 and Parquet export utilities
  - graph_ops.rs: Graph operation types and helpers
  - import.rs: Parquet, directory, and S3 import logic
- Removed conditional #[cfg(feature = "aws")] guards

Related to #94, #95
